chore: 에러 타입을 레이어에 따라 분리
