+++ DevLog/Data/Common/DataLayerError.swift
@@ -1,8 +1,8 @@
 //
-//  Error+.swift
+//  DataLayerError.swift
 //  DevLog
 //
-//  Created by 최윤진 on 11/29/25.
+//  Created by opfic on 3/11/26.
 //
 
 import Foundation
@@ -11,21 +11,6 @@ enum AuthError: Error {
     case notAuthenticated
     case failedToUnlinkLastProvider
     case unsupportedProvider
-}
-
-enum FirestoreError: Error, LocalizedError {
-    case dataNotFound(_ key: String)
-
-    var errorDescription: String? {
-        switch self {
-        case .dataNotFound(let key):
-            return "\(key)가 Firestore에서 존재하지 않음"
-        }
-    }
-}
-
-enum UIError: Error {
-    case notFoundTopViewController
 }
 
 enum DataError: Error {
